introduce spring mvc controller to run ticketService by RESTful request

+++ src/main/java/com/walmart/ticketservice/service/TicketServiceImpl.java
@@ -25,7 +25,8 @@ import com.walmart.ticketservice.database.repository.CustomerRepository;
 import com.walmart.ticketservice.database.repository.SeatHoldRepository;
 import com.walmart.ticketservice.database.repository.SeatOrderRepository;
 import com.walmart.ticketservice.database.repository.VenueRepository;
-import com.walmart.ticketservice.error.TicketServiceException;
+import com.walmart.ticketservice.error.CustomerValidationException;
+import com.walmart.ticketservice.error.SeatHoldNotFoundException;
 import com.walmart.ticketservice.util.ConfirmationCodeUtils;
 
 @Service("ticketService")
@@ -63,17 +64,22 @@ public class TicketServiceImpl implements TicketService {
 
 	@Override
 	public SeatHold findAndHoldSeats(int numSeats, Optional<Integer> minLevel, Optional<Integer> maxLevel, String customerEmail) {
+		if(StringUtils.isEmpty(customerEmail)) {
+			log.warn("Get empty email, so ignore hold request");
+			throw new CustomerValidationException();
+		}
 		LocalDateTime now = LocalDateTime.now();
 		LocalDateTime expired = now.minusSeconds(SEAT_HOLD_EXPIRATION_TIME_IN_SECONDS);
 		Instant expiredInstant = expired.atZone(ZoneId.systemDefault()).toInstant();
 		Instant nowInstant = now.atZone(ZoneId.systemDefault()).toInstant();
 
-		// clean up expired holds
+		// Clean up expired holds
 		List<SeatHold> expiredHolds = seatHoldRepository.findByConfirmationCodeIsNullAndHoldTimeBefore(Date.from(expiredInstant));
 		if(!expiredHolds.isEmpty()) {
 			seatHoldRepository.delete(expiredHolds);
 		}
 
+		// Retreive venue and customer info
 		List<VenueLevel> venueLevels = VenueLevel.getLevels(minLevel, maxLevel);
 		Customer customer = customerRepository.findByEmail(customerEmail);
 		if(null == customer) {
@@ -85,6 +91,7 @@ public class TicketServiceImpl implements TicketService {
 		SeatHold seatHold = new SeatHold();
 		seatHold.setCustomer(customer);
 
+		// Find seats through venue levels
 		int numSeatsToHold = numSeats;
 		for(VenueLevel venueLevel: venueLevels) {
 			if(0 < numSeatsToHold) {
@@ -103,6 +110,7 @@ public class TicketServiceImpl implements TicketService {
 			}
 		}
 
+		// persist the hold
 		if( seatHold.getSeatOrders().isEmpty() ) {
 			log.warn("fail to hold any seat in levels for customer email: ".concat(customerEmail));
 		} else {
@@ -124,19 +132,30 @@ public class TicketServiceImpl implements TicketService {
 		LocalDateTime expired = now.minusSeconds(SEAT_HOLD_EXPIRATION_TIME_IN_SECONDS);
 		Instant expiredInstant = expired.atZone(ZoneId.systemDefault()).toInstant();
 
+		// find SeatHold and verify it
 		SeatHold seatHold = seatHoldRepository.findOne(seatHoldId);
 		if(null == seatHold) {
 			String errorMessage = String.join(": ", "fail on reservation, no SeatHold found. it probably already expired, seatHoldId", String.valueOf(seatHoldId));
 			log.error(errorMessage);
-			throw new TicketServiceException(errorMessage);
+			throw new SeatHoldNotFoundException(errorMessage);
 		}
 		if(seatHold.getHoldTime().before(Date.from(expiredInstant))) {
 			seatHoldRepository.delete(seatHold);
 			String errorMessage = String.join(": ", "fail on reservation, the SeatHold is expired, seatHoldId", String.valueOf(seatHoldId));
 			log.error(errorMessage);
-			throw new TicketServiceException(errorMessage);
+			throw new SeatHoldNotFoundException(errorMessage);
+		}
+		if( !StringUtils.isEmpty(seatHold.getConfirmationCode()) ) {
+			String message = new StringBuilder("The seatHold is already reservated, seatHoldId: ")
+								.append(seatHoldId)
+								.append(", customerEmail: ")
+								.append(customerEmail)
+								.toString();
+			log.warn(message);
+			return seatHold.getConfirmationCode();
 		}
 
+		// verify customer info
 		Customer customer = seatHold.getCustomer();
 		if(null == customer || !StringUtils.equalsIgnoreCase(customer.getEmail(), customerEmail)) {
 			StringBuilder errorMessage = new StringBuilder("Customer Eamil Validation on SeatHold fail, seatHoldId: ")
@@ -144,9 +163,10 @@ public class TicketServiceImpl implements TicketService {
 												.append(", customerEmail: ")
 												.append(customerEmail);
 			log.error(errorMessage.toString());
-			throw new TicketServiceException(errorMessage.toString());
+			throw new CustomerValidationException(errorMessage.toString());
 		}
 
+		// generate confirmation code and check in the hold as reservation
 		String code = ConfirmationCodeUtils.generateCode(seatHoldId, customerEmail);
 		seatHold.setConfirmationCode(code);
 		seatHoldRepository.save(seatHold);
